# Faster version of vtk_grid