Added a faster version of `vtk_grid` which can be used if the topology is known (which is quite common)

Test with a mesh with 100x100 elements:
```jl
julia> @time vtk_grid(edof, Coord, Dof, 4, "vtk_outfile")
  1.652264 seconds (334.70 k allocations: 12.642 MB)

julia> @time vtk_grid(top, Coord, "vtk_outfile")
  0.034052 seconds (79.24 k allocations: 4.011 MB)
```

200x200 elements:
```jl
julia> @time vtk_grid(edof, Coord, Dof, 4, "vtk_outfile")
 25.855713 seconds (1.35 M allocations: 50.621 MB, 0.05% gc time)

julia> @time vtk_grid(top, Coord, "vtk_outfile")
  0.153537 seconds (319.34 k allocations: 15.901 MB)
```


1000x1000 elements (this definitely crashes with the original version):
```jl
julia> @time vtk_grid(top, Coord, "vtk_outfile")
  4.168526 seconds (8.00 M allocations: 376.221 MB, 11.23% gc time)
```