@fredrikekre 2016-03-05
Was thinking about adding meshing for 1D and 3D aswell, and maybe add a type called JuAMesh or something which can be parametrized on the dimension, following the convention for the function spaces, i.e. a "square" mesh in 1, 2 and 3 dimensions.