@KristofferC 2016-03-05
I also thought about something like that. It is a bit annoying to deal with all `Edof, Ex, Ey, Coord, Dof` etc so packing them in a type is definitely something that is needed.

It would also be cool to use `ContMechTensors` (when it is done)` to store for example `Coord` as a Vector of first order tensors.