@fredrikekre 2016-03-05
Yep I can put together a proposed mesh type with meshfunctions, and then enhance it with the use of `ContMechTensors` later on.