@fredrikekre 2016-03-05
Not sure, either it could be just a geometrical object with elements, nodes and coordinates, but then there should be a function that set up a desired number of dofs in each node and generate edof.

Or it could be like now, that you specify the number of dofs when creating the mesh and get back everything.

But technically the dofs doesnt belong to the mesh, idk...