@KristofferC 2016-03-05
My "ideal" situation would be: The mesh is a geometrical concept. It represents nodes, the coordinates of the nodes and how these are connected to elements. Basically coordinates + topology.

Then there is a function that takes a mesh and a collection of field dimensions, like `(3, 1, 1)` which would represent a vector field and two scalar field and from that constructs the Edof.

For example if the field representation would be (3,1) which represent 3 displacements and the pressure a coord would have `x, y, z, p` as dofs and an edof for a triangle would look like:

`x, y, z, x, y, z, x, y, z, p, p ,p` so that each field is close together.