@fredrikekre 2016-03-05
Yes, that makes sense.

Something like:

```jl
mymesh = JuAMesh(startcoordinates,endcoordinates,dim,nElements,...)

edof = create_edof(mymesh,(3,1,1))
```

maybe?