@KristofferC 2016-03-05
Yes something like that. Although I don't think we should return a "naked" edof, it should probably be wrapped in some type as well so you can do `get_dofs(dof_object, i)` to get the dofs of the i:th element.