@fredrikekre 2016-03-05
Yes, and a similar function to get the dofs for a specified field, which can be used for adding the data to the vtk-file. Like:

```jl
displacements = u[get_dofs(dof_object,1)]
pressure = u[get_dofs(dof_object,2)]
vtk_point_data(vtkfile, displacements, "displacement")
vtk_point_data(vtkfile, pressure, "pressure")
```
if `u` is the solution vector, field 1 is displacement and 2 the pressure. But then it is needed to somehow define if its for an element or a field the dofs should be extracted for, else the functions look the same.