@KristofferC 2016-03-05
The function could also look like `get_field(dof_object, 1, u)` which would give the displacements.