@KristofferC 2016-03-05
Good reading to avoid unnecessary commits, like "fix comment": http://gitready.com/advanced/2009/02/10/squashing-commits-with-rebase.html